perf(connections): isolate catalog browser state — kill typing lag
The search input state lived in the ~2k-line AddConnectionDialog, so every
keystroke re-rendered the whole dialog tree (form, About panel, 60 tooltip
cards) — debouncing the scoring wasn't enough. The catalog step (search,
kind chips, results grid + scorers) moves to a self-contained
CatalogBrowser component owning its own state, with memoized cards;
keystrokes now re-render only that subtree.
